src/batch/app/requirements.txt
--extra-index-url https://europe-python.pkg.dev/${PROJECT_NAME}/python-packages/simple/

thg-gcp-utils==1.1.0
pydantic==1.10.7

# Transitive dependencies below
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
Deprecated==1.2.13
google-api-core==2.11.0
google-auth==2.17.2
google-cloud-core==2.3.2
google-cloud-monitoring==2.14.2
google-cloud-storage==2.7.0
google-cloud-trace==1.11.1
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
idna==3.4
importlib-metadata==6.1.0
jaraco.classes==3.2.3
keyring==23.13.1
keyrings.google-artifactregistry-auth==1.1.2
more-itertools==9.1.0
opentelemetry-api==1.16.0
opentelemetry-distro==0.37b0
opentelemetry-exporter-gcp-monitoring==1.4.0a0
opentelemetry-exporter-gcp-trace==1.4.0
opentelemetry-instrumentation==0.37b0
opentelemetry-sdk==1.16.0
opentelemetry-semantic-conventions==0.37b0
pluggy==1.0.0
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyYAML==6.0
requests==2.28.2
rsa==4.9
six==1.16.0
typing_extensions==4.5.0
urllib3==1.26.15
wrapt==1.15.0
zipp==3.15.0
